+++ src/main/java/org/folio/search/configuration/properties/SearchConfigurationProperties.java
@@ -42,6 +42,16 @@ public class SearchConfigurationProperties {
   private long maxBrowseRequestOffset = 500L;
 
   /**
+   * Provides the maximum number of IDs for performing search with batch requests.
+   */
+  private long maxSearchBatchRequestIdsCount = 20_000L;
+
+  /**
+   * Provides the size parameter for querying consortium records (holdings or items).
+   */
+  private int searchConsortiumRecordsPageSize = 5_000;
+
+  /**
    * Provides map with global features configuration. Can be overwritten by tenant configuration.
    */
   private Map<TenantConfiguredFeature, Boolean> searchFeatures = emptyMap();