+++ src/test/java/org/folio/search/utils/TestConstants.java
@@ -37,9 +37,11 @@ public class TestConstants {
   public static final String INVENTORY_BOUND_WITH_TOPIC = "inventory.bound-with";
   public static final String INVENTORY_CLASSIFICATION_TYPE_TOPIC = "inventory.classification-type";
   public static final String CONSORTIUM_INSTANCE_TOPIC = "search.consortium.instance";
-  public static final String BIBFRAME_TOPIC = "search.bibframe";
+  public static final String LINKED_DATA_TOPIC = "linked-data.work";
+  public static final String LINKED_DATA_AUTHORITY_TOPIC = "linked-data.authority";
   public static final String CAMPUS_TOPIC = "inventory.campus";
   public static final String INSTITUTION_TOPIC = "inventory.institution";
+  public static final String LIBRARY_TOPIC = "inventory.library";
 
   public static final String LOCAL_CN_TYPE = "6fd29f52-5c9c-44d0-b529-e9c5eb3a0aba";
   public static final String FOLIO_CN_TYPE = "6e4d7565-b277-4dfa-8b7d-fbf306d9d0cd";
@@ -122,8 +124,12 @@ public class TestConstants {
     return getTopicName(tenantId, INVENTORY_BOUND_WITH_TOPIC);
   }
 
-  public static String bibframeTopic(String tenantId) {
-    return getTopicName(tenantId, BIBFRAME_TOPIC);
+  public static String linkedDataWorkTopic(String tenantId) {
+    return getTopicName(tenantId, LINKED_DATA_TOPIC);
+  }
+
+  public static String linkedDataAuthorityTopic(String tenantId) {
+    return getTopicName(tenantId, LINKED_DATA_AUTHORITY_TOPIC);
   }
 
   public static String inventoryCampusTopic(String tenantId) {
@@ -132,6 +138,10 @@ public class TestConstants {
 
   public static String inventoryInstitutionTopic(String tenantId) {
     return getTopicName(tenantId, INSTITUTION_TOPIC);
+  }
+
+  public static String inventoryLibraryTopic(String tenantId) {
+    return getTopicName(tenantId, LIBRARY_TOPIC);
   }
 
   public static String indexName(String tenantId) {